search-bot/requirements.txt
arxiv
beautifulsoup4
chromadb
duckdb
duckduckgo-search
duckduckgo_search==5.3.1b1
faiss-cpu
huggingface_hub
ipykernel
langchain
langchain-chroma
langchain-community
langchain-groq
langchain-huggingface
langchain-openai
langchain-ollama
langchain-text-splitters
mysql-connector-python
numexpr
openai
pandas
pymupdf
PyPDF2
python-dotenv
pytube
pypdf
sentence_transformers
SQLAlchemy
streamlit
unstructured
validators==0.28.1
wikipedia
youtube_transcript_api
pyodbc 
sqlalchemy
unstructured
pytube
numexpr
gradio